Move httproxy to internal
